+++ xmlschema/validators/globals_.py
@@ -16,9 +16,10 @@ import re
 from ..exceptions import XMLSchemaKeyError, XMLSchemaTypeError, XMLSchemaValueError
 from ..namespaces import XSD_NAMESPACE_PATH, URIDict
 from ..qnames import (
-    get_qname, local_name, XSD_INCLUDE_TAG, XSD_IMPORT_TAG, XSD_REDEFINE_TAG,
-    XSD_NOTATION_TAG, XSD_SIMPLE_TYPE_TAG, XSD_COMPLEX_TYPE_TAG, XSD_GROUP_TAG,
-    XSD_ATTRIBUTE_TAG, XSD_ATTRIBUTE_GROUP_TAG, XSD_ELEMENT_TAG, XSD_ANY_TYPE
+    get_qname, local_name, reference_to_qname, XSD_INCLUDE_TAG, XSD_IMPORT_TAG,
+    XSD_REDEFINE_TAG, XSD_NOTATION_TAG, XSD_SIMPLE_TYPE_TAG, XSD_COMPLEX_TYPE_TAG,
+    XSD_GROUP_TAG, XSD_ATTRIBUTE_TAG, XSD_ATTRIBUTE_GROUP_TAG, XSD_ELEMENT_TAG,
+    XSD_ANY_TYPE
 )
 from .exceptions import XMLSchemaParseError, XMLSchemaNotBuiltError
 from .parseutils import get_xsd_attribute
@@ -372,8 +373,8 @@ class XsdGlobals(XsdBaseComponent):
         self.substitution_groups.clear()
         for xsd_element in self.elements.values():
             if xsd_element.substitution_group:
-                qname = get_qname(xsd_element.target_namespace, xsd_element.substitution_group)
-                if xsd_element.type.name == XSD_ANY_TYPE:
+                qname = reference_to_qname(xsd_element.substitution_group, xsd_element.schema.namespaces)
+                if xsd_element.type.name == XSD_ANY_TYPE and 'type' not in xsd_element.elem.attrib:
                     xsd_element.type = self.elements[qname].type
                 try:
                     self.substitution_groups[qname].add(xsd_element)